requirements.txt
aniso8601==7.0.0
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.7.0
dj-database-url==0.5.0
Django==3.1.7
django-debug-toolbar==3.2
django-filter==2.4.0
django-haystack==3.0
django-heroku==0.3.1
djangorestframework==3.12.2
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
idna==2.5
install==1.3.4
Markdown==3.3.4
oauthlib==3.1.0
Pillow==8.1.1
promise==2.3
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
pysolr==3.9.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
Rx==1.6.1
singledispatch==3.6.1
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.3
whitenoise==5.2.0
